feat!: remove legacy hook support
