+++ lib/arke/group.ex
@@ -20,7 +20,6 @@ defmodule Arke.System.Group do
       Module.register_attribute(__MODULE__, :parameters, accumulate: true, persist: false)
       Module.register_attribute(__MODULE__, :system_group, accumulate: false, persist: true)
       Module.register_attribute(__MODULE__, :arke_hooks, accumulate: true)
-      Module.register_attribute(__MODULE__, :arke_legacy_warning, accumulate: false)
       Module.put_attribute(__MODULE__, :system_group, true)
 
       import unquote(__MODULE__),
@@ -28,25 +27,13 @@ defmodule Arke.System.Group do
 
       import Arke.Hook.DSL
 
-      @before_compile {Arke.Hook.Compiler, :__before_compile_group__}
+      @before_compile Arke.Hook.DSL
 
       def group_from_attr(),
         do: Keyword.get(__MODULE__.__info__(:attributes), :group, []) |> List.first()
 
       def is_group?(),
         do: Keyword.get(__MODULE__.__info__(:attributes), :system_group, []) |> List.first()
-
-      def on_unit_load(arke, data, _persistence_fn), do: {:ok, data}
-      def before_unit_load(_arke, data, _persistence_fn), do: {:ok, data}
-      def on_unit_validate(_arke, unit), do: {:ok, unit}
-      def before_unit_validate(_arke, unit), do: {:ok, unit}
-      def on_unit_struct_encode(unit, _), do: {:ok, unit}
-
-      defoverridable on_unit_load: 3,
-                     before_unit_load: 3,
-                     on_unit_validate: 2,
-                     before_unit_validate: 2,
-                     on_unit_struct_encode: 2
     end
   end
