+++ test/support/hook_arke.ex
@@ -111,110 +111,3 @@ defmodule Arke.Test.HookArke do
     :ok
   end
 end
-
-defmodule Arke.Test.LegacyHookArke do
-  @moduledoc """
-  Support arke keeping the pre-pipeline callback heads, to prove the
-  deprecation shim registers them as pipeline entries.
-  """
-  use Arke.System
-
-  @arke_legacy_warning false
-
-  arke id: :legacy_hook_arke do
-    parameter(:hook_flag, :string, required: false)
-  end
-
-  def before_create(_arke, unit) do
-    send(self(), {:legacy, :before_create})
-
-    case Arke.Core.Unit.get_value(unit, :hook_flag) do
-      "fail_before_create" -> {:error, [%{context: "legacy_test", message: "failed on purpose"}]}
-      "stamp" -> {:ok, Arke.Core.Unit.update(unit, hook_flag: "stamped")}
-      _ -> {:ok, unit}
-    end
-  end
-
-  def on_create(_arke, unit) do
-    send(self(), {:legacy, :on_create})
-
-    case Arke.Core.Unit.get_value(unit, :hook_flag) do
-      "fail_on_create" -> {:error, [%{context: "legacy_test", message: "on_create failed"}]}
-      _ -> {:ok, unit}
-    end
-  end
-
-  def before_update(_arke, old_unit, unit) do
-    send(self(), {:legacy, :before_update, old_unit.id})
-    {:ok, unit}
-  end
-
-  def on_update(_arke, _old_unit, unit) do
-    send(self(), {:legacy, :on_update})
-    {:ok, unit}
-  end
-
-  def before_delete(_arke, unit) do
-    send(self(), {:legacy, :before_delete})
-    {:ok, unit}
-  end
-
-  def on_delete(_arke, unit) do
-    send(self(), {:legacy, :on_delete})
-    {:ok, unit}
-  end
-
-  def on_struct_encode(_arke, _unit, data, _opts) do
-    {:ok, Map.put(data, :legacy_encoded, true)}
-  end
-
-  def register() do
-    [] =
-      Arke.handle_manager(
-        [arke_from_attr() |> Map.update!(:id, &to_string/1)],
-        :arke_system,
-        :arke
-      )
-
-    :ok
-  end
-end
-
-defmodule Arke.Test.LegacyHookGroup do
-  @moduledoc """
-  Support group keeping the pre-pipeline `*_unit_*` callback heads, to prove
-  the group shim. Bound to `legacy_hook_arke`.
-  """
-  use Arke.System.Group
-
-  @arke_legacy_warning false
-
-  group id: :legacy_hook_group do
-  end
-
-  def before_unit_create(_arke, unit) do
-    send(self(), {:legacy_group, :before_unit_create})
-    {:ok, unit}
-  end
-
-  def on_unit_create(_arke, unit) do
-    send(self(), {:legacy_group, :on_unit_create})
-    {:ok, unit}
-  end
-
-  def on_unit_delete(_arke, unit) do
-    send(self(), {:legacy_group, :on_unit_delete})
-    {:ok, unit}
-  end
-
-  def register() do
-    [] =
-      Arke.handle_manager(
-        [%{id: "legacy_hook_group", label: "Legacy hook group", arke_list: ["legacy_hook_arke"]}],
-        :arke_system,
-        :group
-      )
-
-    :ok
-  end
-end
