+++ test/test_helper.exs
@@ -8,6 +8,4 @@ Arke.Test.Bootstrap.start()
 Arke.Test.CreateArke.support_parameter()
 Arke.Test.CreateArke.support_arke()
 Arke.Test.HookArke.register()
-Arke.Test.LegacyHookArke.register()
-Arke.Test.LegacyHookGroup.register()
 Arke.Test.Sandbox.checkpoint()
